+++ src/commonMain/kotlin/com/hoc081098/flowext/NullValue.kt
@@ -5,8 +5,7 @@ package com.hoc081098.flowext
  * This allows for writing faster generic code instead of using `Option`.
  * This is only used as an optimisation technique in low-level code.
  */
-@Suppress("ClassName")
-public object NULL_Value {
+public object NullValue {
   @Suppress("UNCHECKED_CAST", "NOTHING_TO_INLINE")
   public inline fun <T> unbox(v: Any?): T = if (this === v) null as T else v as T
 }