+++ src/commonMain/kotlin/com/hoc081098/flowext/retryWhenWithDelayStrategy.kt
@@ -1,3 +1,27 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2021 Petrus Nguyễn Thái Học
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package com.hoc081098.flowext
 
 import kotlinx.coroutines.delay
@@ -6,17 +30,6 @@ import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.retryWhen
 import kotlin.math.pow
 import kotlin.time.Duration
-
-public fun <T> Flow<T>.retryWhenWithDelayStrategy(
-  strategy: DelayStrategy,
-  predicate: suspend FlowCollector<T>.(cause: Throwable, attempt: Long) -> Boolean
-): Flow<T> = retryWhen { cause, attempt ->
-  predicate(cause, attempt).also {
-    if (it) {
-      delay(strategy.duration(cause, attempt))
-    }
-  }
-}
 
 public fun interface DelayStrategy {
   public fun duration(cause: Throwable, attempt: Long): Duration
@@ -37,6 +50,17 @@ public fun interface DelayStrategy {
     override fun duration(cause: Throwable, attempt: Long): Duration =
       (if (attempt <= 0L) initialDelay else initialDelay * factor.pow(attempt.toDouble()))
         .coerceAtMost(maxDelay)
+  }
+}
+
+public fun <T> Flow<T>.retryWhenWithDelayStrategy(
+  strategy: DelayStrategy,
+  predicate: suspend FlowCollector<T>.(cause: Throwable, attempt: Long) -> Boolean
+): Flow<T> = retryWhen { cause, attempt ->
+  predicate(cause, attempt).also {
+    if (it) {
+      delay(strategy.duration(cause, attempt))
+    }
   }
 }