+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 0.2.0-SNAPSHOT
 
 -   Update
-    -   `Kotlin` to `1.6.0`.
-    -   `KotlinX Coroutines` to `1.6.0-RC2`.
+    -   `Kotlin` to `1.6.10`.
+    -   `KotlinX Coroutines` to `1.6.0`.
 
 -   Do not propagate cancellation to the upstream in Flow `flatMapFirst` operators
     (Related to https://github.com/Kotlin/kotlinx.coroutines/pull/2964).